+++ modules/Sources/Features/SyncProgress/SyncProgressView.swift
@@ -13,20 +13,21 @@ import UIComponents
 import Models
 
 public struct SyncProgressView: View {
-    var store: SyncProgressStore
+    @Perception.Bindable var store: StoreOf<SyncProgress>
     
-    public init(store: SyncProgressStore) {
+    public init(store: StoreOf<SyncProgress>) {
         self.store = store
     }
     
     public var body: some View {
-        WithViewStore(store, observe: { $0 }) { viewStore in
+        WithPerceptionTracking {
             VStack(spacing: 5) {
-                if viewStore.isSyncing {
+                if store.isSyncing {
                     HStack {
-                        Text(viewStore.syncStatusMessage)
+                        Text(store.syncStatusMessage)
                             .font(.custom(FontFamily.Inter.regular.name, size: 10))
-
+                            .foregroundColor(Asset.Colors.primary.color)
+                        
                         // Frame height 0 is expected value because we want SwiftUI to ignore it
                         // for the vertical placement computation.
                         ProgressView()
@@ -34,21 +35,35 @@ public struct SyncProgressView: View {
                             .frame(width: 11, height: 0)
                     }
                 } else {
-                    Text(viewStore.syncStatusMessage)
-                        .multilineTextAlignment(.center)
-                        .font(.custom(FontFamily.Inter.regular.name, size: 10))
-                        .padding(.horizontal, 35)
+                    if let errorMessage = store.lastKnownErrorMessage {
+                        Button {
+                            store.send(.errorMessageTapped)
+                        } label: {
+                            Text(store.syncStatusMessage)
+                                .multilineTextAlignment(.center)
+                                .font(.custom(FontFamily.Inter.regular.name, size: 10))
+                                .padding(.horizontal, 35)
+                                .foregroundColor(Asset.Colors.primary.color)
+                        }
+                    } else {
+                        Text(store.syncStatusMessage)
+                            .multilineTextAlignment(.center)
+                            .font(.custom(FontFamily.Inter.regular.name, size: 10))
+                            .padding(.horizontal, 35)
+                            .foregroundColor(Asset.Colors.primary.color)
+                    }
                 }
                 
-                Text(String(format: "%0.1f%%", viewStore.syncingPercentage * 100))
+                Text(String(format: "%0.1f%%", store.syncingPercentage * 100))
                     .font(.custom(FontFamily.Inter.black.name, size: 10))
                     .foregroundColor(Asset.Colors.primary.color)
                 
-                ProgressView(value: viewStore.syncingPercentage, total: 1.0)
+                ProgressView(value: store.syncingPercentage, total: 1.0)
                     .progressViewStyle(ZashiSyncingProgressStyle())
             }
-            .onAppear { viewStore.send(.onAppear) }
-            .onDisappear { viewStore.send(.onDisappear) }
+            .onAppear { store.send(.onAppear) }
+            .onDisappear { store.send(.onDisappear) }
+            .alert($store.scope(state: \.alert, action: \.alert))
         }
     }
 }
@@ -56,15 +71,34 @@ public struct SyncProgressView: View {
 #Preview {
     SyncProgressView(
         store:
-            SyncProgressStore(
+            StoreOf<SyncProgress>(
                 initialState: .init(
                     lastKnownSyncPercentage: Float(0.43),
                     synchronizerStatusSnapshot: SyncStatusSnapshot(.syncing(0.41)),
                     syncStatusMessage: "Syncing"
                 )
             ) {
-                SyncProgressReducer()
+                SyncProgress()
             }
     )
     .background(.red)
+}
+
+// MARK: - Store
+
+extension SyncProgress {
+    public static var initial = StoreOf<SyncProgress>(
+        initialState: .initial
+    ) {
+        SyncProgress()
+    }
+}
+
+// MARK: - Placeholders
+
+extension SyncProgress.State {
+    public static let initial = SyncProgress.State(
+        lastKnownSyncPercentage: 0,
+        synchronizerStatusSnapshot: .initial
+    )
 }