+++ CHANGELOG.md
@@ -9,6 +9,7 @@ directly impact users rather than highlighting other crucial architectural updat
 ### Added
 - Expanded transaction lists all text memos.
 - Biometric lock is used to protect Delete Zashi, Export Private Data and Send features.
+- Tapping on the error message label in the sync progress shows an alert view with the details of the error.
 
 ### Fixed
 - Sometimes, Zashi crashed when the shield button was tapped. We fixed the crash, but shielding won't be possible due to funds being below the threshold.